Refactor OssLicenseDetail components into detail package
