+++ ui-wear-compose-material3/src/debug/java/com/github/droibit/oss_licenses/ui/wear/compose/material3/internal/detail/OssLicenseDetailScreenPreview.kt
@@ -1,4 +1,4 @@
-package com.github.droibit.oss_licenses.ui.wear.compose.material3.internal
+package com.github.droibit.oss_licenses.ui.wear.compose.material3.internal.detail
 
 import androidx.compose.runtime.Composable
 import androidx.wear.compose.material3.MaterialTheme
