+++ ui-wear-compose-material3/src/main/java/com/github/droibit/oss_licenses/ui/wear/compose/material3/internal/detail/OssLicenseDetailScreen.kt
@@ -1,4 +1,4 @@
-package com.github.droibit.oss_licenses.ui.wear.compose.material3.internal
+package com.github.droibit.oss_licenses.ui.wear.compose.material3.internal.detail
 
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.lazy.LazyListState
