+++ ui-wear-compose-material3/src/main/java/com/github/droibit/oss_licenses/ui/wear/compose/material3/internal/list/OssLicenseListScreen.kt
@@ -43,4 +43,3 @@ internal fun OssLicenseListScreen(
     )
   }
 }
-
